+++ launch/ackermann_mux_launch.py
@@ -14,6 +14,7 @@
 # limitations under the License.
 
 # Author: Brighten Lee
+#         Hongrui Zheng
 
 import os
 from ament_index_python.packages import get_package_share_directory
@@ -24,11 +25,11 @@ from launch_ros.actions import Node
 
 
 def generate_launch_description():
-    default_config_locks = os.path.join(get_package_share_directory('twist_mux'),
-                                        'config', 'twist_mux_locks.yaml')
-    default_config_topics = os.path.join(get_package_share_directory('twist_mux'),
-                                         'config', 'twist_mux_topics.yaml')
-    default_config_joystick = os.path.join(get_package_share_directory('twist_mux'),
+    default_config_locks = os.path.join(get_package_share_directory('ackermann_mux'),
+                                        'config', 'ackermann_mux_locks.yaml')
+    default_config_topics = os.path.join(get_package_share_directory('ackermann_mux'),
+                                         'config', 'ackermann_mux_topics.yaml')
+    default_config_joystick = os.path.join(get_package_share_directory('ackermann_mux'),
                                            'config', 'joystick.yaml')
 
     return LaunchDescription([